backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
yt-dlp==2024.12.13
deepgram-sdk==2.12.0
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
googletrans==4.0.0rc1
ffmpeg-python==0.2.0 
